chore: created AppWidget for better initialization
